STEP 9. Break up fragment hydration.
