+++ altitude/static/js/fragments/person-name-editor.js
@@ -1,0 +1,10 @@
+import { focusFragmentElement } from "./helpers.js"
+
+export function hydratePersonNameEditorFragment({ fragmentEl }) {
+    focusFragmentElement(
+        fragmentEl,
+        fragmentEl.dataset.appAutofocusSelector,
+        fragmentEl.dataset.appSelectOnFocus === "true",
+    )
+}
+
