+++ altitude/static/js/frontend-app.js
@@ -10,7 +10,8 @@ import {
     dragMoveListener,
     setFixedPositionWhileDragging,
 } from "./common/dragon-drop.js"
-import { closeModal, showAssetDetailModal, showModal } from "./common/modal.js"
+import { showAssetDetailModal } from "./common/modal.js"
+import { hydrateAppFragments } from "./fragments/index.js"
 import { setImgSrcAndWait } from "./search-results/detail-navigator.js"
 import "./search-results/dragon-drop.js"
 
@@ -840,215 +841,7 @@ export class FrontendApp {
     }
 
     hydrateFragments(root) {
-        this.findFragmentRoots(root, "person-name-editor").forEach(
-            (fragmentEl) => {
-                this.hydratePersonNameEditorFragment(fragmentEl)
-            },
-        )
-
-        this.findFragmentRoots(root, "modal").forEach((fragmentEl) => {
-            this.hydrateModalFragment(fragmentEl)
-        })
-
-        this.findFragmentRoots(root, "image-detail").forEach((fragmentEl) => {
-            this.hydrateImageDetailFragment(fragmentEl)
-        })
-
-        this.findFragmentRoots(root, "search-results").forEach((fragmentEl) => {
-            this.hydrateSearchResultsFragment(fragmentEl)
-        })
-    }
-
-    findFragmentRoots(root, fragmentName) {
-        if (!(root instanceof Element || root instanceof Document)) {
-            return []
-        }
-
-        const selector = `[data-app-fragment="${fragmentName}"]`
-        const roots = []
-
-        if (root instanceof Element && root.matches(selector)) {
-            roots.push(root)
-        }
-
-        roots.push(...root.querySelectorAll(selector))
-
-        return roots
-    }
-
-    hydrateModalFragment(fragmentEl) {
-        showModal({
-            minWidthPx: fragmentEl.dataset.appModalMinWidth,
-            title: fragmentEl.dataset.appModalTitle,
-        })
-
-        this.initializeModalFragment(fragmentEl)
-        this.focusFragmentElement(
-            fragmentEl,
-            fragmentEl.dataset.appModalAutofocusSelector,
-            fragmentEl.dataset.appModalSelectOnFocus === "true",
-        )
-        this.bindModalFragment(fragmentEl)
-    }
-
-    initializeModalFragment(fragmentEl) {
-        if (fragmentEl.dataset.appModalKind === "view-settings") {
-            this.initializeViewSettingsModalFragment(fragmentEl)
-        }
-    }
-
-    initializeViewSettingsModalFragment(fragmentEl) {
-        const showFields = this.context.getGridMetadataFields()
-
-        fragmentEl
-            .querySelectorAll('input[type="checkbox"]')
-            .forEach((checkboxEl) => {
-                checkboxEl.checked = showFields.has(checkboxEl.value)
-            })
-
-        if (fragmentEl.dataset.appViewSettingsBound === "true") {
-            return
-        }
-
-        fragmentEl.dataset.appViewSettingsBound = "true"
-
-        fragmentEl.addEventListener("change", (event) => {
-            const checkboxEl = event.target
-            if (
-                !(checkboxEl instanceof HTMLInputElement) ||
-                checkboxEl.type !== "checkbox"
-            ) {
-                return
-            }
-
-            const fieldName = checkboxEl.value
-            const checked = checkboxEl.checked
-
-            if (checked) {
-                this.context.addGridMetadataField(fieldName)
-            } else {
-                this.context.removeGridMetadataField(fieldName)
-            }
-
-            this.dispatch(Const.events.viewSettingChanged, {
-                fieldName,
-                checked,
-            })
-        })
-    }
-
-    focusFragmentElement(fragmentEl, selector, selectOnFocus = false) {
-        if (!selector) {
-            return
-        }
-
-        const focusEl = fragmentEl.matches(selector)
-            ? fragmentEl
-            : fragmentEl.querySelector(selector)
-
-        if (!focusEl) {
-            return
-        }
-
-        focusEl.focus()
-
-        if (selectOnFocus && typeof focusEl.select === "function") {
-            focusEl.select()
-        }
-    }
-
-    bindModalFragment(fragmentEl) {
-        if (fragmentEl.dataset.appModalBound === "true") {
-            return
-        }
-
-        fragmentEl.dataset.appModalBound = "true"
-
-        fragmentEl.addEventListener("htmx:afterRequest", (event) => {
-            if (event.detail.successful !== true) {
-                return
-            }
-
-            this.handleModalFragmentSuccess(fragmentEl, event)
-        })
-    }
-
-    handleModalFragmentSuccess(fragmentEl, event) {
-        this.dispatchModalFragmentSuccessEvent(fragmentEl, event)
-        this.runModalFragmentSuccessAction(fragmentEl)
-
-        if (fragmentEl.dataset.appModalCloseOnSuccess !== "false") {
-            closeModal()
-        }
-    }
-
-    dispatchModalFragmentSuccessEvent(fragmentEl, event) {
-        const eventKey = fragmentEl.dataset.appModalSuccessEvent
-        if (!eventKey) {
-            return
-        }
-
-        const eventName = Const.events[eventKey]
-        if (!eventName) {
-            console.warn(`Unknown modal success event key: ${eventKey}`)
-            return
-        }
-
-        this.dispatch(
-            eventName,
-            this.buildModalFragmentSuccessDetail(fragmentEl, event),
-        )
-    }
-
-    buildModalFragmentSuccessDetail(fragmentEl, event) {
-        return {
-            ...this.parseModalFragmentDetail(
-                fragmentEl.dataset.appModalSuccessDetail,
-            ),
-            ...this.parseModalFragmentTargetDetail(fragmentEl, event),
-        }
-    }
-
-    parseModalFragmentTargetDetail(fragmentEl, event) {
-        const detail = {}
-
-        Object.entries(fragmentEl.dataset).forEach(([key, value]) => {
-            const prefix = "appModalSuccessDetailTargetAttr"
-            if (!key.startsWith(prefix)) {
-                return
-            }
-
-            const detailKey = this.datasetKeySuffixToDetailKey(
-                key.slice(prefix.length),
-            )
-
-            detail[detailKey] =
-                event.target?.getAttribute?.(value) ?? detail[detailKey]
-        })
-
-        return detail
-    }
-
-    datasetKeySuffixToDetailKey(keySuffix) {
-        if (!keySuffix) {
-            return ""
-        }
-
-        return keySuffix.charAt(0).toLowerCase() + keySuffix.slice(1)
-    }
-
-    runModalFragmentSuccessAction(fragmentEl) {
-        const action = fragmentEl.dataset.appModalSuccessAction
-        if (!action) {
-            return
-        }
-
-        if (action === "close-folder-context-menu") {
-            this.closeFolderContextMenu(fragmentEl.dataset.appModalFolderId)
-            return
-        }
-
-        console.warn(`Unknown modal success action: ${action}`)
+        hydrateAppFragments({ root, app: this })
     }
 
     closeFolderContextMenu(folderId) {
@@ -1063,60 +856,6 @@ export class FrontendApp {
         }
     }
 
-    parseModalFragmentDetail(jsonValue) {
-        if (!jsonValue) {
-            return {}
-        }
-
-        try {
-            return JSON.parse(jsonValue)
-        } catch (error) {
-            console.error("Error parsing modal success detail", error)
-            return {}
-        }
-    }
-
-    hydratePersonNameEditorFragment(fragmentEl) {
-        this.focusFragmentElement(
-            fragmentEl,
-            fragmentEl.dataset.appAutofocusSelector,
-            fragmentEl.dataset.appSelectOnFocus === "true",
-        )
-    }
-
-    hydrateImageDetailFragment(fragmentEl) {
-        if (fragmentEl.dataset.appImageDetailBound === "true") {
-            return
-        }
-
-        fragmentEl.dataset.appImageDetailBound = "true"
-
-        const imgEl = fragmentEl.querySelector("img")
-        if (!imgEl) {
-            return
-        }
-
-        this.Alpine.store(Const.state.imageDetailLoading).value = true
-
-        setImgSrcAndWait(imgEl, fragmentEl.dataset.appImageDetailUrl)
-            .then(() => {
-                showAssetDetailModal({
-                    title: fragmentEl.dataset.appImageDetailTitle,
-                    width: Number(fragmentEl.dataset.appImageDetailWidth),
-                    height: Number(fragmentEl.dataset.appImageDetailHeight),
-                })
-
-                this.dispatch(Const.events.detailShown, {
-                    assetId: fragmentEl.dataset.appImageDetailAssetId,
-                })
-            })
-            .catch((error) => {
-                console.error("Error loading asset image", error)
-            })
-            .finally(() => {
-                this.Alpine.store(Const.state.imageDetailLoading).value = false
-            })
-    }
 
     hydrateSearchResultsFragment(fragmentEl) {
         const resultsTotal = Number(fragmentEl.dataset.resultsTotal || 0)
